남은 정적 분석 지적을 판단해서 처리한다 (65 -> 2건)
남은 65건을 하나씩 봤다. 고칠 수 있으면 고치고, 구조적으로 불가능한 자리는 이유를 적어
NOLINT 했다. 이유 없는 억제는 남기지 않았다.

찾은 실제 결함:

- Component 생성자가 가상 함수를 불러 기본값이 틀린 타입으로 적용됐다. 생성자가 initialize() ->
  가상 getTypeInfo() 를 부르는데, 생성 중에는 객체가 아직 Component 라 파생 타입이 아니라 기반
  타입의 TypeInfo 가 나온다. MeshComponent 를 만들어도 "Component" 이름으로 기본값을 찾았다.
  실제 생성 경로는 타입을 아는 쪽이 이미 올바르게 넘겨 준다 — GameObject::addComponent<T> 와
  GameObjectManager 의 이름 기반 생성이 둘 다 applyTypeDefaults(파생 TypeInfo) 를 부른다. 즉
  생성자 호출은 중복이면서 틀린 조회였고 컴포넌트를 만들 때마다 헛일을 했다. 호출과 (호출자가
  없어진) initialize() 를 걷어냈다. initialize() 는 private 이고 호출자가 생성자뿐이며, 기본값
  데이터는 게임이 setPath 를 부르지 않으면 로드되지 않는다(Resource/ 에 <Defaults> 노드도 없다).

- FrameRendererPassExecute 의 SSAO PSO 폴백이 자기 자신이었다.
  getEnginePso(SSAO) != 0 ? getEnginePso(SSAO) : getEnginePso(SSAO) — 참·거짓이 같아 아무 효과가
  없고 함수만 세 번 불렀다. 형제 패스는 전부 다른 PSO 로 폴백한다(DepthPrepass->Shadow,
  GBufferAlbedo->GBuffer, Tonemap->Present). 복사하면서 대체 대상을 바꾸지 않은 자리다. 폴백을
  짐작해 넣지 않고, PSO 가 0 이면 drawFullscreen 이 건너뛰므로 형제들처럼 그대로 넘긴다.

- StringBuilder::appendFormat 이 재시도 루프에서 매번 std::forward 했다. 버퍼가 모자라면 같은
  인자로 다시 포맷하는데 그때는 이미 이동된 값을 쓴다. formatstring 은 값을 읽어 찍기만 하므로
  lvalue 로 넘겨 위험을 없앴다.

- 컴포넌트 붙여넣기가 실패를 삼켰다. pasteComponentAsNew 가 역직렬화 결과를 계산해 놓고 아무도
  읽지 않았다. 둘 다 실패해도 빈 컴포넌트를 붙이고 실행 취소 항목까지 남겼다. 경고를 남긴다.

- Material 의 죽은 코드. 첫 순회가 packSize 를 계산했지만 뒤따르는 if 는 본문이 주석뿐인 빈
  블록이었고 값은 아무도 읽지 않았다. 실제 패킹은 두 번째 순회가 다시 계산해서 한다.

- 콘텐츠 브라우저 필터 라벨이 string_view 였다. 쓰는 쪽 셋이 모두 .data() 를 ImGui 로 넘기는데
  ImGui 는 널 종단을 요구한다. 표가 전부 리터럴이라 우연히 맞을 뿐이라 타입을 const utf8* 로 바꿔
  계약을 적었다.

- 소멸자·생성자의 가상 호출 12곳을 클래스 이름으로 한정했다. 파괴 중에는 파생 재정의가 이미
  사라진 뒤라 이 클래스의 것이 불린다 — 지금 동작이 의도한 것이므로 코드로 적었다.

- 확장자 길이를 손으로 쓰던 자리(.bin/.xml 은 -4, .json 은 -5)를 표로 돌게 했다. Json/Xml
  직렬화의 같은 본문 두 분기는 조건으로 합쳤다. ShaderBaker 의 SPIR-V 두 케이스는 묶었다.
  남은 int 곱셈 후 확대(SPIR-V 파서, 두 곳의 reserve 힌트)도 size_t 로 맞췄다.

오탐이라 이유만 남긴 것: Base{std::move(other)} 뒤의 파생 멤버 읽기, 타입 이름·## 인자를 받는
매크로, 순서가 규약인 분기, 크기를 함께 넘기는 append(data(), count), float3/float4x4 를 연속
float 로 훑는 자리(배치는 옆의 static_assert 가 보장한다), 뮤텍스 락의 noexcept 지적.

.clang-tidy 에서 Padding 검사를 끈다. 걸린 셋이 전부 인스턴스가 하나뿐인 매니저이거나 20행짜리
정적 표라 아끼는 양이 무의미한데, 이 저장소는 "생성자 초기화는 선언 순서" 규약이라 멤버를 옮기면
초기화 목록도 같이 옮겨야 한다.

검증: Debug·Shipping·ASan 빌드 경고 0, lint 6/6, Shipping nogpu 5/5, EngineTest 423/423,
CoreTest 168/168, EditorTest 34/34, SmokeTest 19/19, ReflectionTest 100(+1 스킵),
네 백엔드 × 에디터 [Error] 0건, 패널 덤프 14창/빈 0개.
마지막 clang-tidy 재실행은 중간에 중단해 최종 수치(2건 -> 0건 예상)는 확인하지 못했다.

+++ Source/Engine/Graphics/Shader/Reflection/ShaderReflectionSpirv.cpp
@@ -187,7 +187,7 @@ namespace sw
             {
                 const uint32 target = pWords[offset + 1];
                 const utf8*  pStr   = reinterpret_cast<const utf8*>( &pWords[offset + 2] );
-                const size_t maxLen = ( instrWords - 2 ) * 4;
+                const size_t maxLen = static_cast<size_t>( instrWords - 2 ) * 4;
                 mapName[target]     = string( pStr, strnlen( pStr, maxLen ) );
             }
             else if ( opcode == ShaderReflectionSpirvInternal::kOpMemberName && instrWords >= 4 )
@@ -195,7 +195,7 @@ namespace sw
                 const uint32 target                = pWords[offset + 1];
                 const uint32 memberIndex           = pWords[offset + 2];
                 const utf8*  pStr                  = reinterpret_cast<const utf8*>( &pWords[offset + 3] );
-                const size_t maxLen                = ( instrWords - 3 ) * 4;
+                const size_t maxLen                = static_cast<size_t>( instrWords - 3 ) * 4;
                 mapMemberName[target][memberIndex] = string( pStr, strnlen( pStr, maxLen ) );
             }
             else if ( opcode == ShaderReflectionSpirvInternal::kOpDecorate && instrWords >= 3 )
